Refactor creature creator sections into modular files
