+++ src/apps/library/create/creature/index.ts
@@ -1,6 +1,8 @@
 // src/apps/library/create/creature/index.ts
 export { CreateCreatureModal } from "./modal";
-export { mountCoreStatsSection } from "./section-core-stats";
+export { mountCreatureBasicsSection } from "./section-basics";
+export { mountCreatureStatsAndSkillsSection } from "./section-stats-and-skills";
+export { mountCreatureSensesAndDefensesSection } from "./section-senses-and-defenses";
 export { mountEntriesSection } from "./section-entries";
 export { mountSpellsKnownSection } from "./section-spells-known";
 export * from "./presets";
